+++ src/main/java/brig/concord/psi/impl/delegate/ConcordYamlDelegateFactory.java
@@ -2,7 +2,7 @@ package brig.concord.psi.impl.delegate;
 
 import brig.concord.meta.ConcordMetaTypeProvider;
 import brig.concord.meta.model.LoopArrayItemMetaType;
-import brig.concord.meta.model.call.CallInParamMetaType;
+import brig.concord.meta.model.call.CallInParamsMetaType;
 import brig.concord.meta.model.call.CallMetaType;
 import brig.concord.psi.impl.yaml.YAMLQuotedTextImpl_;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
@@ -52,8 +52,8 @@ public class ConcordYamlDelegateFactory {
 
     private static PsiNamedElement createKeyValueDelegate(YAMLKeyValue keyValue) {
         ConcordMetaTypeProvider instance = ConcordMetaTypeProvider.getInstance(keyValue.getProject());
-        YamlMetaType valueMetaType = instance.getResolvedKeyValueMetaTypeMeta(keyValue);
-        if (valueMetaType instanceof CallInParamMetaType) {
+        YamlMetaType valueMetaType = instance.getResolvedMetaType(keyValue);
+        if (valueMetaType instanceof CallInParamsMetaType) {
             return new YamlFlowCallInParamDelegate(keyValue);
         }
         return null;