+++ src/main/java/brig/concord/psi/ref/CallInParamDefinitionReference.java
@@ -3,6 +3,9 @@ package brig.concord.psi.ref;
 import brig.concord.completion.provider.FlowCallParamsProvider;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.psi.*;
+import com.intellij.psi.util.CachedValueProvider;
+import com.intellij.psi.util.CachedValuesManager;
+import com.intellij.psi.util.PsiModificationTracker;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.yaml.psi.YAMLKeyValue;
 
@@ -14,12 +17,36 @@ public class CallInParamDefinitionReference extends PsiReferenceBase.Poly<YAMLKe
 
     @Override
     public ResolveResult @NotNull [] multiResolve(boolean incompleteCode) {
-        PsiElement inParamDefinition = FlowCallParamsProvider.getInstance().inParamDefinition(getElement());
-        if (inParamDefinition == null) {
+        PsiElement inParamDef = CachedValuesManager.getCachedValue(getElement(), new InParamsDefinitionCachedValueProvider(getElement()));
+
+        if (inParamDef == null) {
             return ResolveResult.EMPTY_ARRAY;
         }
 
-        return PsiElementResolveResult.createResults(inParamDefinition);
+        return PsiElementResolveResult.createResults(inParamDef);
+    }
+
+    private static final class InParamsDefinitionCachedValueProvider implements CachedValueProvider<PsiElement> {
+
+        private final YAMLKeyValue inParamElement;
+
+        private InParamsDefinitionCachedValueProvider(YAMLKeyValue inParamElement) {
+            this.inParamElement = inParamElement;
+        }
+
+        @Override
+        public Result<PsiElement> compute() {
+            PsiElement callDefinition = FlowCallParamsProvider.findCallKv(inParamElement);
+            if (callDefinition != null) {
+                return CachedValueProvider.Result.create(resolveInner(), PsiModificationTracker.MODIFICATION_COUNT, callDefinition);
+            } else {
+                return CachedValueProvider.Result.create(null, PsiModificationTracker.MODIFICATION_COUNT, inParamElement);
+            }
+        }
+
+        private PsiElement resolveInner() {
+            return FlowCallParamsProvider.getInstance().inParamDefinition(inParamElement);
+        }
     }
 }