+++ src/main/java/brig/concord/psi/ref/FlowDefinitionReference.java
@@ -4,6 +4,9 @@ import brig.concord.psi.ProcessDefinition;
 import brig.concord.psi.ProcessDefinitionProvider;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.psi.*;
+import com.intellij.psi.util.CachedValueProvider;
+import com.intellij.psi.util.CachedValuesManager;
+import com.intellij.psi.util.PsiModificationTracker;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.yaml.psi.YAMLScalar;
 
@@ -15,18 +18,35 @@ public class FlowDefinitionReference extends PsiReferenceBase.Poly<YAMLScalar> i
 
     @Override
     public ResolveResult @NotNull [] multiResolve(boolean incompleteCode) {
-        ProcessDefinition process = ProcessDefinitionProvider.getInstance().get(getElement());
-        if (process == null) {
-            return ResolveResult.EMPTY_ARRAY;
-        }
-
-        String flowName = getElement().getTextValue();
-        PsiElement flowDef = process.flow(flowName);
+        PsiElement flowDef = CachedValuesManager.getCachedValue(getElement(), new FlowDefinitionCachedValueProvider(getElement()));
         if (flowDef == null) {
             return ResolveResult.EMPTY_ARRAY;
         }
-
         return PsiElementResolveResult.createResults(flowDef);
+    }
+
+    private static final class FlowDefinitionCachedValueProvider implements CachedValueProvider<PsiElement> {
+
+        private final YAMLScalar callElement;
+
+        private FlowDefinitionCachedValueProvider(YAMLScalar callElement) {
+            this.callElement = callElement;
+        }
+
+        @Override
+        public Result<PsiElement> compute() {
+            return CachedValueProvider.Result.create(resolveInner(), PsiModificationTracker.MODIFICATION_COUNT, callElement);
+        }
+
+        private PsiElement resolveInner() {
+            ProcessDefinition process = ProcessDefinitionProvider.getInstance().get(callElement);
+            if (process == null) {
+                return null;
+            }
+
+            String flowName = callElement.getTextValue();
+            return process.flow(flowName);
+        }
     }
 }